requirements.txt
streamlit
sentence-transformers
faiss-cpu
pdfplumber
google-generativeai
torch
